# api/openapi-spec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A full-text and semantic search engine for exploring multilingual news headlines and articles",
        "title": "Tusker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@tusker.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://opensource.org/licenses/Apache-2.0"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/articles/search": {
            "get": {
                "description": "Simple text search with automatic field selection and weighting. Cacheable and bookmarkable. Application determines optimal search strategy based on index configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Simple query string search",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"climate change\"",
                        "description": "SearchStringQuery query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "Results per page (default: 100, max: 10000)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor (base64-encoded from previous response)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"english\"",
                        "description": "SearchStringQuery language: english, serbian (default: english)",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SearchStringQuery results with pagination metadata",
                        "schema": {
                            "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - missing or invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/articles/semantic_search": {
            "get": {
                "description": "Perform vector-based semantic search using the embedding model. Supports pagination via cursor. Application determines optimal search strategy based on index configuration. Results are cacheable and bookmarkable.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Vector-based semantic search",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"climate change\"",
                        "description": "SearchStringQuery query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "Results per page (default: 100, max: 10000)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor (base64-encoded from previous response)",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SearchStringQuery results with pagination metadata",
                        "schema": {
                            "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.SemanticSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - missing or invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/articles/structured": {
            "post": {
                "description": "Execute structured search queries with explicit control over fields, weights, and operators. Supports match and multi_match query types. Follows Elasticsearch query DSL pattern.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Structured search API",
                "parameters": [
                    {
                        "description": "Structured search request with query wrapper",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SearchStringQuery results with pagination metadata",
                        "schema": {
                            "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid query structure or parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "501": {
                        "description": "Query type not supported by searcher backend",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/capabilities": {
            "get": {
                "description": "Returns the search paradigms exposed by the running backend. Useful for clients to discover available capabilities (e.g. whether semantic search is enabled).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "capabilities"
                ],
                "summary": "Report supported search paradigms",
                "responses": {
                    "200": {
                        "description": "Supported search paradigms",
                        "schema": {
                            "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_types_query.Capabilities"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.Article": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.ArticleMetadata"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string",
                    "format": "string"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.ArticleMetadata": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Content metadata",
                    "type": "string"
                },
                "importedAt": {
                    "description": "System metadata",
                    "type": "string"
                },
                "publishedAt": {
                    "type": "string"
                },
                "sourceId": {
                    "description": "Essential source tracking",
                    "type": "string"
                },
                "sourceName": {
                    "type": "string"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.ArticleSearchResult": {
            "type": "object",
            "properties": {
                "article": {
                    "description": "Embedded Article struct for search results",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.Article"
                        }
                    ]
                },
                "score": {
                    "description": "Score rank between 0 and 1",
                    "type": "number"
                },
                "score_normalized": {
                    "description": "ScoreNormalized is the normalized(between 0-1) score",
                    "type": "number"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.BooleanParams": {
            "type": "object",
            "required": [
                "expression"
            ],
            "properties": {
                "expression": {
                    "type": "string",
                    "minLength": 1
                },
                "language": {
                    "type": "string"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.HybridParams": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "k": {
                    "type": "integer",
                    "minimum": 1
                },
                "language": {
                    "type": "string"
                },
                "query": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.MatchParams": {
            "type": "object",
            "required": [
                "field",
                "query"
            ],
            "properties": {
                "field": {
                    "type": "string"
                },
                "fuzziness": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "query": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.MultiMatchParams": {
            "type": "object",
            "required": [
                "fields",
                "query"
            ],
            "properties": {
                "field_weights": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "number",
                        "format": "float64"
                    }
                },
                "fields": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "language": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "query": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.PhraseParams": {
            "type": "object",
            "required": [
                "fields",
                "query"
            ],
            "properties": {
                "fields": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "language": {
                    "type": "string"
                },
                "query": {
                    "type": "string",
                    "minLength": 1
                },
                "slop": {
                    "type": "integer",
                    "maximum": 3,
                    "minimum": 0
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.QueryWrapper": {
            "type": "object",
            "properties": {
                "boolean": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.BooleanParams"
                },
                "hybrid": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.HybridParams"
                },
                "match": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.MatchParams"
                },
                "multi_match": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.MultiMatchParams"
                },
                "phrase": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.PhraseParams"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.SearchRequest": {
            "type": "object",
            "properties": {
                "cursor": {
                    "type": "string"
                },
                "query": {
                    "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.QueryWrapper"
                },
                "size": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.SearchResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "hits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.ArticleSearchResult"
                    }
                },
                "max_score": {
                    "type": "number"
                },
                "next_cursor": {
                    "type": "string"
                },
                "page_max_score": {
                    "type": "number"
                },
                "total_matches": {
                    "type": "integer"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_api_dto.SemanticSearchResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "hits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_DjordjeVuckovic_tusker_internal_api_dto.Article"
                    }
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "github_com_DjordjeVuckovic_tusker_internal_types_query.Capabilities": {
            "type": "object",
            "properties": {
                "boolean": {
                    "type": "boolean"
                },
                "match": {
                    "type": "boolean"
                },
                "multi_match": {
                    "type": "boolean"
                },
                "phrase": {
                    "type": "boolean"
                },
                "semantic": {
                    "type": "boolean"
                },
                "string_query": {
                    "type": "boolean"
                }
            }
        }
    }
}